requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
python-multipart
pydantic-settings
sqlalchemy
pytest
httpx